+++ docs_src/src/guide/changelog.md
@@ -12,6 +12,12 @@ meta:
 
 [[toc]]
 
+## v3.7.0
+
+- `*goyave.Response` now implements `http.Hijacker`. Note that status handlers and middleware (requests finalization step in their lifecycle) will still work for hijacked connections.
+- Static file serving will no longer print "no such file or directory" to the error logger.
+- Static file serving optimization: check file existence once instead of twice.
+
 ## v3.6.0
 
 - Set content type to `application/json; charset=utf-8` instead of `application/json` when using `response.JSON()`.